requirements.txt
customtkinter
pyautogui
pillow
requests